app/requirements.txt
Flask>=2.0.1
tensorflow>=2.17.0
numpy>=1.20
Pillow>=8.3.1
prometheus-flask-exporter>=0.18.4
Werkzeug>=2.0.3
protobuf>=3.20.3